Revert "Merge remote-tracking branch 'origin/patch-1' into patch-1"
This reverts commit 9f49f98e3c381020c10fad82655bb1a3816737c3.

+++ modules/core/include/opencv2/core/hal/intrin_cpp.hpp
@@ -614,7 +614,7 @@ template<typename _Tp, int n>
 inline v_reg<typename V_TypeTraits<_Tp>::abs_type, n> v_popcount(const v_reg<_Tp, n>& a)
 {
     v_reg<typename V_TypeTraits<_Tp>::abs_type, n> b = v_reg<typename V_TypeTraits<_Tp>::abs_type, n>::zero();
-    for (int i = 0; i < (int)(n*sizeof(_Tp)); i++)
+    for( int i = 0; i < n*sizeof(_Tp); i++ )
         b.s[i/sizeof(_Tp)] += popCountTable[v_reinterpret_as_u8(a).s[i]];
     return b;
 }
